Fix gulpfile issue with build:html
